requirements.txt
amqp      ==     2.3.2
anyjson   ==        0.3.3
attrs     ==        18.2.0
Automat   ==       0.7.0
billiard  ==       3.5.0.4
certifi   ==        2018.10.15
chardet   ==        3.0.4
constantly ==       15.1.0
Django     ==       2.1.3
django-filebrowser==3.11.1
django-grappelli  ==2.12.1
django-guardian   ==1.4.1
django-userena    ==2.0.1
djangorestframework==3.9.0
easy-thumbnails   ==2.5
future            ==0.17.1
html2text         ==2018.1.9
hyperlink         ==18.0.0
idna              ==2.7
incremental       ==17.5.0
Markdown          ==3.0.1
mysqlclient       ==1.3.13
numpy             ==1.15.4
numpy-quaternion  ==2018.11.6.10.44.9
Pillow            ==5.3.0
pip               ==18.1
PyHamcrest        ==1.9.0
python3-memcached ==1.51
pytz              ==2018.7
qiniu             ==7.2.2
requests          ==2.20.1
setuptools        ==40.5.0
simplejson        ==3.16.0
six               ==1.11.0
Twisted           ==18.9.0
urllib3           ==1.24.1
vine              ==1.1.4
zope.interface    ==4.6.0

